[9.1] A Simple Diffuse Material
- Added utility functions to generate random vectors.
This will simulate the random light rays bouncing off the
sphere.
- Determine if ray is on the correct hemisphere by comparing
against the surface normal
- Limiting number of child rays.
The ray_color function is recursive. When will it stop recursing?
When it fails to hit anything. In some cases, however, that may
be a long time — long enough to blow the stack.
To guard against that, limited the maximum recursion depth,
returning no light contribution at the maximum depth.

%SOFTWARE